Added ability to download ops manager ova

+++ deploy.sh
@@ -1,18 +1,28 @@
 #!/usr/bin/env bash
 [[ -z $DEBUGX ]] || set -x
 
+repo_root="$(cd "$(dirname "$0")" && pwd)"
+
 ## Always call finish when scripts exits (error or success)
 function finish {
-  if [ -d "$temp_dir" ]; then
+  if [ -d "$temp_dir" ] && [ "$CLEANUP" = true ]; then
     rm -rf "$temp_dir"
   fi
 }
 trap finish EXIT
 
+if ls -d $repo_root/tmp.*/ &>/dev/null; then
+  #use existing tmp dir
+  temp_dir="$(ls -d $repo_root/tmp.*)"
+else
+  # create temporary work dir
+  temp_dir="$(mktemp -d -p $repo_root)"
+fi
+
 #################################################
 ### DEFINE VARIABLES (WITH OPTIONAL DEFAULTS) ###
 #################################################
-MODE=
+CLEANUP=${CLEANUP:-true}
 
 ########################
 ### DEFINE FUNCTIONS ###
@@ -24,6 +34,8 @@ function usage() {
   echo "Environment variables to be set:"
   echo "PIVNET_API_TOKEN - API Token to access Tanzu network downloads"
   echo "PLATFORM_AUTOMATION_VERSION - The version of platform automation to use, if not set then latest will be selected"
+  echo "OPSMAN_VERSION - The version of ops manager to deploy, if not set then latest will be selected"
+  echo "CLEANUP - Automatically cleans up downloaded artefacts, defaults to true"
   echo ""
   echo "Command line arguments:"
   echo "-h - Script help/usage"
@@ -35,6 +47,11 @@ function usage() {
 function validate() {
   
   [ -z "$PIVNET_API_TOKEN" ] && return 1
+
+  if [[ ! $CLEANUP =~ ^(true|false)$ ]]; then 
+    echo "Error: CLEANUP env var must be true or false"
+    return 1
+  fi
 
   return 0
 }
@@ -64,6 +81,39 @@ function prepare_docker() {
 
 }
 
+docker_run() {
+  docker run \
+    --volume="${repo_root}:/workdir" \
+    --volume="${temp_dir}:/tempdir" \
+    --workdir="/workdir" \
+    "platform-automation-image:${PLATFORM_AUTOMATION_VERSION}" \
+    "$@"
+}
+
+function deploy_opsman() {
+
+  # get the latest ops manager version if explicit version not set
+  [ -z "$OPSMAN_VERSION" ] && OPSMAN_VERSION=`curl -s https://network.tanzu.vmware.com/api/v2/products/ops-manager/releases | jq -r '.releases | first | .version'`
+
+  if [ ! -f ${temp_dir}/ops-manager-vsphere-$OPSMAN_VERSION.ova ]; then
+  
+    echo "Downloading ops manager version: $OPSMAN_VERSION"
+    docker_run om download-product \
+    --pivnet-api-token="$PIVNET_API_TOKEN" \
+    --pivnet-product-slug=ops-manager \
+    --product-version="$OPSMAN_VERSION" \
+    --file-glob='ops-manager-vsphere-*.ova' \
+    --output-directory="/tempdir"
+
+  else
+    echo "Skipping ops manager download as version: $OPSMAN_VERSION is already present" 
+  fi
+
+
+
+
+}
+
 #############################################
 ### Read parameters from the command line ###
 #############################################
@@ -77,9 +127,10 @@ done
 ## Main script logic
 function main() {
 
-  temp_dir="$(mktemp -d -p `pwd`)"
+
 
   prepare_docker
+  deploy_opsman
   
 }
